add exception and move httpservice
