+++ AggregatorMicroService/Exceptions/ModelException.cs
@@ -1,0 +1,8 @@
+﻿namespace AggregatorMicroService.Exceptions;
+
+public class ModelException : Exception
+{
+    public ModelException(string message) : base(message)
+    {
+    }
+}
