+++ AggregatorMicroService/Middlewares/ExceptionHandlingMiddleware.cs
@@ -32,6 +32,7 @@ public class ExceptionHandlingMiddleware
             UnauthorizedAccessException => StatusCodes.Status401Unauthorized,
             ForbiddenException => StatusCodes.Status403Forbidden,
             NotFoundException => StatusCodes.Status404NotFound,
+            ModelException=> StatusCodes.Status422UnprocessableEntity,
             _ => StatusCodes.Status500InternalServerError
         };
