+++ AggregatorMicroService/Service/HttpService.cs
@@ -1,9 +1,10 @@
 ﻿using System.Net;
 using System.Text;
 using AggregatorMicroService.Exceptions;
+using AggregatorMicroService.Service.Abstraction;
 using Newtonsoft.Json;
 
-namespace AggregatorMicroService.HttpClient;
+namespace AggregatorMicroService.Service;
 
 public class HttpService : IHttpService
 {
@@ -62,6 +63,7 @@ public class HttpService : IHttpService
                 HttpStatusCode.Unauthorized => new UnauthorizedAccessException(responseMessage.ReasonPhrase),
                 HttpStatusCode.Forbidden => new ForbiddenException(responseMessage.ReasonPhrase),
                 HttpStatusCode.NotFound => new NotFoundException(responseMessage.ReasonPhrase),
+                HttpStatusCode.UnprocessableEntity => new ModelException(responseMessage.ReasonPhrase);
                 _ => new Exception(responseMessage.ReasonPhrase)
             };
     }
